+++ src/lcm1602/lcm1602.c
@@ -135,9 +135,9 @@ lcm1602_context lcm1602_i2c_init(int bus, int address, bool is_expander,
     return dev;
 }
 
-lcm1602_context lcm1602_gpio_init(uint8_t rs, uint8_t enable,
-                                  uint8_t d0, uint8_t d1, uint8_t d2,
-                                  uint8_t d3, uint8_t num_columns,
+lcm1602_context lcm1602_gpio_init(int rs, int enable,
+                                  int d0, int d1, int d2,
+                                  int d3, uint8_t num_columns,
                                   uint8_t num_rows)
 {
     lcm1602_context dev =
